WASM: single-pass merged rendering for all surface categories
Refactor BrainRenderer with dual render-target pattern (clearing +
preserving) and single merged VBO/IBO for all WASM surface categories.

Multi-pass compositing via PreserveColorContents proved unreliable on
WebGL — only one surface category was visible at a time. Fix by merging
brain, source space, digitizer, and BEM/sensor surfaces into a single
draw call per frame.

Changes:
- BrainRenderer: add ensureRenderTargets() with dual-RT pair, track
  colorTexture pointer for rebuild, add hasMergedContent() guard
- BrainRenderer: beginFrame/beginPreservingPass/endPass API for
  multi-pass (retained for future use when Qt fixes the bug)
- BrainView (WASM): merge all visible surfaces into one 'default'
  group; single drawMergedSurfaces() in clearing pass
- BrainView: setAutoRenderTarget(false), use custom render targets
- Fix QWARN -> qWarning in test_mne_projections, test_mne_forward_ops
- Fix format specifier %d -> %lld for qsizetype in mne_label_ssp

+++ src/tools/inverse/mne_label_ssp/main.cpp
@@ -207,7 +207,7 @@ int main(int argc, char *argv[])
             continue;
         }
 
-        fprintf(stderr, "  Matched %d gain-matrix columns\n", colIndices.size());
+        fprintf(stderr, "  Matched %lld gain-matrix columns\n", static_cast<long long>(colIndices.size()));
 
         //
         // Extract restricted gain matrix columns
@@ -256,7 +256,7 @@ int main(int argc, char *argv[])
     //
     // Write projectors to output file
     //
-    fprintf(stderr, "Writing %d projector(s) to %s...\n", projectors.size(), outFile.toUtf8().constData());
+    fprintf(stderr, "Writing %lld projector(s) to %s...\n", static_cast<long long>(projectors.size()), outFile.toUtf8().constData());
 
     QFile outIO(outFile);
     FiffStream::SPtr stream = FiffStream::start_file(outIO);
